% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_details_hmap.R
\name{extract_details_hmap}
\alias{extract_details_hmap}
\title{Get voucher info from hmap records}
\usage{
extract_details_hmap(html.file)
}
\arguments{
\item{abc}{abc}
}
\value{
a
}
\description{
A
}
