% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_urls_flickr.R
\name{get_urls_flickr}
\alias{get_urls_flickr}
\title{Search for image locations on flickr}
\usage{
get_urls_flickr(reptileData, subset, flickrAPI)
}
\arguments{
\item{reptileData}{Reptile database data from the read_reptile_data() function}

\item{subset}{Required. A family subset to work with, can be a single family.}

\item{flickrAPI}{Required. An api key provided by flickr.}
}
\value{
A dataframe of images resulting from the search
}
\description{
A wrapper to search through species for search_flickr. Will use the tags c("snake", "reptile", "tortoise", "turtle", "lizard", "crocodile", "alligator",
"crocodilian", "amphisbaenian", "tuatara", "gharial", "caiman")
}
