% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_urls_hmap.R
\name{get_urls_hmap}
\alias{get_urls_hmap}
\title{Search for image locations on hmap}
\usage{
get_urls_hmap(reptileData, subset)
}
\arguments{
\item{reptileData}{Reptile database data from the read_reptile_data()
function}

\item{subset}{Required. For herpmapper this has to be an Order from
c("Amphisbaenia", "Crocodylia", "Lacertilia", "Serpentes", "Sphenodon",
"Testudines")}
}
\value{
A dataframe of image information
}
\description{
A function to find and pull out image locations for HerpMapper.
Unlike other functions for sources this one requires an order by order
approach used for Herpmapper.
}
