% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_urls_wiki.R
\name{get_urls_wiki}
\alias{get_urls_wiki}
\title{Search for image locations on wikimedia}
\usage{
get_urls_wiki(reptileData, subset)
}
\arguments{
\item{reptileData}{Reptile database data from the read_reptile_data() function}

\item{subset}{Required. A family subset to work with, can be a single family.}
}
\value{
A dataframe of images resulting from the search
}
\description{
A wrapper for the search_wiki function to loop through many species.
}
