% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_flickr.R
\name{search_flickr}
\alias{search_flickr}
\title{Get info from flickr searches}
\usage{
search_flickr(
  search,
  api_key,
  tags,
  tag_m,
  content_type,
  extra_call,
  page_call
)
}
\arguments{
\item{search}{search text string}

\item{api_key}{Required. flickr API key}

\item{tags}{e.g., c("snake", "reptile", "tortoise", "turtle", "lizard", "crocodile", "alligator",
"crocodilian", "amphisbaenian", "tuatara", "gharial", "caiman")}

\item{tag_m}{"any"}

\item{content_type}{"1"}

\item{extra_call}{"date_taken,geo,license"}

\item{page_call}{"250"}
}
\value{
A list of two dataframes, one of the meta information, one of the image information.
}
\description{
A function to call the flickr API to run a search for images,
then a second call to get the static image URLs.
}
