% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_wiki.R
\name{search_wiki}
\alias{search_wiki}
\title{Get info from wiki searches}
\usage{
search_wiki(searchterm = NA)
}
\arguments{
\item{searchterm}{Species search term}
}
\value{
A dataframe of wikimedia results.
}
\description{
A function to search wikimedia for images.
}
